requirements.txt
matplotlib==3.7.2
numpy==1.24.4
pandas==2.0.3
Pillow==10.0.0
Pillow==10.0.0
scikit_learn==1.3.0
torch==2.0.0+cu118
torchinfo==1.8.0
torchvision==0.15.1+cu118
